ARM: shmobile: armadillo800eva: fix clock inversion
When operating in left-justfied mode both the frame-clock and the
bit-clock need to be inverted to be standards compliant.

This means that the exta clock inversion setting in the armadillo800eva
machine driver for CPU component should now be removed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
Acked-by: Simon Horman <horms+renesas@verge.net.au>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 947a37cd38796f5b196a934353165a001cbcb0a9)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>